requirements.txt
pandas
pytest
pylint
pyscreenshot
pydocstyle
numpy
matplotlib
scipy
opencv-python
pillow
pyinstaller
pymongo
pyqt5
pytesseract
virtualbox
pyyaml
requests
configobj
xlrd
pyscreenshot
lmfit
jupyter